set PetImage foreign key onUpdate/onDelete action to CASCADE

+++ src/main/java/panda/demo/action/entity/pet/PetBulkDeleteExAction.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import panda.demo.entity.Pet;
 import panda.demo.entity.query.PetImageQuery;
+import panda.lang.Systems;
 import panda.lang.mutable.MutableInt;
 import panda.mvc.annotation.At;
 
@@ -18,15 +19,19 @@ public class PetBulkDeleteExAction extends PetBulkDeleteAction {
 
 	@Override
 	protected void deleteDataList(List<Pet> dataList, MutableInt count) {
-		List<Long> ids = new ArrayList<Long>(dataList.size());
-		for (Pet f : dataList) {
-			ids.add(f.getId());
+		if (Systems.IS_OS_APPENGINE) {
+			// GAE does not support foreign key CASCADE action
+			// we need delete PetImage manually
+			List<Long> ids = new ArrayList<Long>(dataList.size());
+			for (Pet f : dataList) {
+				ids.add(f.getId());
+			}
+	
+			PetImageQuery piq = new PetImageQuery();
+			piq.pid().in(ids);
+			getDao().deletes(piq);
 		}
-
-		PetImageQuery piq = new PetImageQuery();
-		piq.pid().in(ids);
-		getDao().deletes(piq);
-
+		
 		super.deleteDataList(dataList, count);
 	}
 }
